Move tests for `replace` into separate file
